requirements.txt
# requirements.txt
beautifulsoup4==4.12.2
pydantic
fastapi
python-dotenv
router
asyncpg
aiofiles
aiohttp
#numpy
#matplotlib
pytz
requests
uvicorn
sqlalchemy
alembic
psycopg2-binary
web3
aiofiles
packaging
solc-select
slither-analyzer
selenium
fake_useragent
lxml
pyyaml
eth-brownie
selenium-wire
